# docs/openapi/openapi-v1.yaml
openapi: 3.0.3
info:
  title: TechDengue — API v1
  version: 1.0.0
  description: |
    API **Edital-Core++** para indicadores EPI, operação de campo, evidências e relatórios.
    Segurança via **OAuth2/OIDC** (Bearer JWT). Todas as rotas v1 retornam `application/json`
    por padrão, exceto downloads (PDF/CSV).
servers:
  - url: https://api.techdengue.local/v1
    description: Ambiente de homologação (exemplo)
  - url: http://localhost:8080/v1
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Atividade:
      type: object
      required: [municipio_cod_ibge, status]
      properties:
        id: { type: string, format: uuid }
        origem: { type: string, enum: [DENUNCIA, PLANO, ALERTA] }
        municipio_cod_ibge: { type: string }
        bairro: { type: string }
        equipe: { type: string }
        sla_deadline: { type: string, format: date-time }
        status: { type: string, enum: [CRIADA, EM_ANDAMENTO, ENCERRADA] }
        criado_em: { type: string, format: date-time }
        atualizado_em: { type: string, format: date-time }
...
    UploadInfo:
      type: object
      properties:
        carga_id: { type: string }
        status: { type: string, enum: [RECEBIDO, VALIDANDO, ERRO, CARREGADO] }
        qualidade_url: { type: string }
        created_at: { type: string, format: date-time }
...
paths:
...
  /atividades:
    get:
      tags: [Campo]
      summary: Lista atividades
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [CRIADA, EM_ANDAMENTO, ENCERRADA] }
        - in: query
          name: municipio_cod_ibge
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Atividade" }
    post:
      tags: [Campo]
      summary: Cria atividade
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Atividade" }
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Atividade" }
  /atividades/{id}:
    get:
      tags: [Campo]
      summary: Detalhe de atividade
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Atividade" }
        "404": { description: Não encontrado }
    patch:
      tags: [Campo]
      summary: Atualiza parcialmente uma atividade
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Atividade" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Atividade" }
...
  /etl/sinan/import:
    post:
      tags: [ETL]
      summary: Import CSV SINAN Arboviroses (PoC - dataset de exemplo)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                arquivo: { type: string, format: binary }
      responses:
        "202":
          description: Recebido e enfileirado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadInfo" }
  /etl/liraa/import:
    post:
      tags: [ETL]
      summary: Import CSV LIRAa (IIP/IBR) (PoC - dataset de exemplo)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                arquivo: { type: string, format: binary }
      responses:
        "202":
          description: Recebido e enfileirado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadInfo" }
...